feat:Simple RBAC system
Create a simple RBAC system that will give user a jwt token in login, the token will contains the role(currently either admin or user). Then send the token in head as bearer token. Which will be decoded and verified and if the role is admin for test endpoint show users the result.