fix(main): Fix the login details going in as query rather than form as it was supposed to
In the login endpoint, there was some problem and was sending the credentials as a query parameter rather than as form data. It was due to not defining the username and password as form, instead it was written as username:Annotated[str,Form()]. It is now changed as username:str=Form().