feat(signup): Add option for user to signup
Users can now use /user end point to signup and the info provided will be stored in the backend. Also the password is hashed. Currently, the user when login will only have role of user by default and can be changed in the database. TODO: Create another protected endpoint to change role of users.